requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.34.0
colorama==0.4.6
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2.3
django-compat==1.0.15
django-cors-headers==4.3.1
django-crontab==0.7.1
django-decouple==2.1
djangorestframework==3.14.0
freezegun==1.2.2
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
isodate==0.6.1
kombu==5.3.3
oauthlib==3.2.2
packaging==23.1
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.23.4
#psycopg2==2.9.7
#psycopg2-binary==2.9.7
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.7.0
pyparsing==3.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
schedule==1.2.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==6.4.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
vine==5.1.0
wcwidth==0.2.9
whitenoise==5.3.0
